refactor: share CORS helpers
